hf_ocr_space/requirements.txt
fastapi
uvicorn
rapidocr_onnxruntime==1.2.3
opencv-python-headless
numpy
pydantic
opencc-python-reimplemented
